Complete refactoring of VimTestUtils
