+++ Src/VimTestUtils/Utilities/StaContext.cs
@@ -1,14 +1,7 @@
 ﻿using System;
-using System.Collections.Generic;
-using System.Collections.ObjectModel;
-using System.Collections.Concurrent;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
-using System.Threading;
-using Vim.Extensions;
-using System.Windows.Threading;
 using System.Diagnostics;
+using System.Threading;
+using System.Windows.Threading;
 
 namespace Vim.UnitTest.Utilities
 {
