+++ Src/VimTestUtils/Utilities/StaTestFrameworkTypeDiscoverer.cs
@@ -1,16 +1,6 @@
 ﻿using System;
-using System.Collections.Generic;
-using System.Collections.ObjectModel;
-using System.Collections.Concurrent;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
-using System.Threading;
-using Vim.Extensions;
-using System.Windows.Threading;
-using System.Diagnostics;
-using Xunit.Sdk;
 using Xunit.Abstractions;
+using Xunit.Sdk;
 
 namespace Vim.UnitTest.Utilities
 {
