+++ Src/VimTestUtils/VimTestUtils.csproj
@@ -8,11 +8,6 @@
     <TargetFramework>net472</TargetFramework>
   </PropertyGroup>
   <ItemGroup>
-    <Compile Remove="Exports\**" />
-    <EmbeddedResource Remove="Exports\**" />
-    <None Remove="Exports\**" />
-  </ItemGroup>
-  <ItemGroup>
     <Reference Include="PresentationCore" />
     <Reference Include="PresentationFramework" />
     <Reference Include="System" />
@@ -33,8 +28,5 @@
     <PackageReference Include="Moq" Version="4.5.28" />
     <PackageReference Include="xunit" Version="2.4.1" />
     <PackageReference Include="xunit.extensibility.execution" Version="2.4.1" />
-  </ItemGroup>
-  <ItemGroup>
-    <ProjectReference Include="..\VimCore\VimCore.fsproj" />
   </ItemGroup>
 </Project>
