+++ Test/VimCoreTest/ChannelTest.cs
@@ -3,6 +3,7 @@ using Microsoft.VisualStudio.Text.Tagging;
 using Xunit;
 using Vim;
 using Vim.Extensions;
+using Vim.EditorHost;
 
 namespace Vim.UnitTest
 {
