+++ Test/VimCoreTest/HistorySessionTest.cs
@@ -2,6 +2,7 @@
 using System.Collections.Generic;
 using System.Linq;
 using System.Text;
+using Vim.EditorHost;
 using Xunit;
 
 namespace Vim.UnitTest
