+++ Test/VimCoreTest/KeyInputTest.cs
@@ -1,5 +1,6 @@
 ﻿using System;
 using System.Linq;
+using Vim.EditorHost;
 using Vim.Extensions;
 using Xunit;
