+++ Test/VimCoreTest/MultiSelectionIntegrationTest.cs
@@ -1,6 +1,7 @@
 ﻿using System.Linq;
 using Microsoft.VisualStudio.Text;
 using Microsoft.VisualStudio.Text.Editor;
+using Vim.EditorHost;
 using Vim.UnitTest;
 using Vim.UnitTest.Mock;
 using Xunit;
