+++ Test/VimCoreTest/RegisterValueTest.cs
@@ -3,6 +3,7 @@ using System.Collections.Generic;
 using System.Linq;
 using System.Text;
 using Microsoft.FSharp.Collections;
+using Vim.EditorHost;
 using Xunit;
 
 namespace Vim.UnitTest
