+++ Test/VimCoreTest/VisualModeTest.cs
@@ -5,6 +5,7 @@ using Moq;
 using Xunit;
 using Vim.Modes.Visual;
 using Vim.UnitTest.Mock;
+using Vim.EditorHost;
 
 namespace Vim.UnitTest
 {
