+++ Test/VimWpfTest/VimMouseProcessorTest.cs
@@ -4,6 +4,7 @@ using System;
 using System.Collections.Generic;
 using System.Linq;
 using System.Text;
+using Vim.EditorHost;
 using Vim.UI.Wpf.Implementation.Mouse;
 using Vim.UnitTest;
 using Xunit;
